requirements.txt
# 必要なパッケージを1行ずつ記載してください
streamlit>=1.38
openai>=1.40
python-docx>=1.1.0

pymupdf
pdfplumber


# --- OCR（画像・スキャンPDF用） ---
pytesseract>=0.3.10
Pillow>=10.0.0     # pytesseractに必要

# --- 補助（Unicode正規化など） ---
unicodedata2>=15.1.0  # macOSでの結合文字(例: モデル)対策

pandas>=2.2.2
openai>=1.51.0
pypandoc>=1.13; sys_platform != 'win32'  # pandocがある場合
docx2pdf>=0.1.8; sys_platform != 'linux' # macOS/Windowsで有効

streamlit[pdf]==1.38.0
pypdfium2>=4.30.0

xlsxwriter>=3.2.0        # Excel出力 (.xlsx)

openpyxl>=3.1.0

regex
reportlab

janome


plotly
matplotlib
numpy
kaleido

# --- 実行環境補足 ---
# macOSでは以下を brew でシステムにインストール
#   brew install tesseract
#   brew install tesseract-lang tesseract-lang-jpn